backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Database
pymongo==4.10.1

# Data Validation
pydantic==2.9.2
pydantic-settings==2.6.0

# PDF Processing
pdfplumber==0.11.4
pytesseract==0.3.13
pdf2image==1.17.0
Pillow==10.4.0

# AI/ML
google-generativeai==0.8.3

# Configuration
python-dotenv==1.0.1
python-multipart==0.0.12

# Caching
redis==5.2.0

# HTTP Client
requests==2.32.3

# Monitoring
sentry-sdk[fastapi]==2.17.0

# Testing
pytest==8.3.3
pytest-mock==3.14.0
pytest-asyncio==0.24.0
httpx==0.27.2
